AAC in Cantonese for Aphasia using LetMeTalk
I have been researching suitable app to use for my father who is
suffering severe aphasia due to stroke and alzhemeir.

Apart from one recommeded from the speech therapist I also research
another one more suitable to test with my father when I had a chance.
At the mean time I wish to post here in the GitHub with the hope to for
others to give their experience in using the app for elderly with
aphasia problem.

The profile is a collection of pictures and the associated title which
linked with a suitable text to speech (tts, in this case the EKHO tts
developed by Cameron Wong from eguidedog.org.hk) for use in the app
namely LetMeTalk, which is popular and can be downloaded.

I will write down the procedure of installing the app and associated tts
in separate document.